+++ src/index.ts
@@ -10,7 +10,7 @@ export { EventTimer } from './event-timer';
 export {
 	bypassCommercialMetricsSampling,
 	initCommercialMetrics,
-} from './sendCommercialMetrics';
+} from './send-commercial-metrics';
 export type { ThirdPartyTag } from './types';
 export { adSizes, getAdSize } from './ad-sizes';
 export type { SizeKeys, AdSizeString, AdSize } from './ad-sizes';
