+++ src/send-commercial-metrics.spec.ts
@@ -3,8 +3,8 @@ import {
 	_,
 	bypassCommercialMetricsSampling,
 	initCommercialMetrics,
-} from './sendCommercialMetrics';
-import type { Metric, Property, TimedEvent } from './sendCommercialMetrics';
+} from './send-commercial-metrics';
+import type { Metric, Property, TimedEvent } from './send-commercial-metrics';
 
 const {
 	Endpoints,
